// src/api/generated/model/importJobDetailsResponse.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ImportJobDetailsResponseStatus } from './importJobDetailsResponseStatus';
import type { ImportJobDetailsResponseType } from './importJobDetailsResponseType';
import type { ImportJobErrorDto } from './importJobErrorDto';

export interface ImportJobDetailsResponse {
  importJobId: string;
  type: ImportJobDetailsResponseType;
  status: ImportJobDetailsResponseStatus;
  fileName: string;
  readCount: number;
  writeCount: number;
  skipCount: number;
  startedAt: string;
  finishedAt: string;
  errorMessage: string;
  failures?: ImportJobErrorDto[];
}