// src/api/generated/model/import-job-details-response.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ImportJobDetailsResponseStatus } from './import-job-details-response-status';
import type { ImportJobDetailsResponseType } from './import-job-details-response-type';
import type { ImportJobErrorDto } from './import-job-error-dto';

export interface ImportJobDetailsResponse {
  importJobId: string;
  type: ImportJobDetailsResponseType;
  status: ImportJobDetailsResponseStatus;
  fileName: string;
  readCount: number;
  writeCount: number;
  skipCount: number;
  startedAt: string;
  finishedAt: string;
  errorMessage: string;
  failures?: ImportJobErrorDto[];
}